Bug 3453  - Fix and Code Changes for loggers
