+++ api/CcsSso.Core.ReportingScheduler/Program.cs
@@ -81,9 +81,9 @@ namespace CcsSso.Core.ReportingScheduler
             services.AddSingleton<ICSVConverter, CSVConverter>();
             services.AddSingleton<IFileUploadToCloud, FileUploadToCloud>();
             services.AddHostedService<OrganisationReportingJob>();
-            //services.AddHostedService<UserReportingJob>();
-            //services.AddHostedService<ContactReportingJob>();
-            //services.AddHostedService<AuditReportingJob>();
+            services.AddHostedService<UserReportingJob>();
+            services.AddHostedService<ContactReportingJob>();
+            services.AddHostedService<AuditReportingJob>();
           });
     }
