+++ api/CcsSso.Shared/Domain/Constants/Constant.cs
@@ -26,4 +26,76 @@ namespace CcsSso.Shared.Domain.Constants
     public const int EmaildomainnameNameMaxCharacters = 63;
     public const int EmaildomainMaxCharacters = 63;
   }
+
+  public static class UserHeaderMap
+  {
+    public const string ID = "id";
+    public const string UserName = "username";
+    public const string OrganisationID = "organisation_id";
+    public const string FirstName = "firstname";
+    public const string LastName = "lastname";
+    public const string Title = "title";
+    public const string mfaEnabled = "mfa_enabled";
+    public const string AccountVerified = "accountverified";
+    public const string SendUserRegistrationEmail = "send_user_registrationemail";
+    public const string IsAdminUser = "is_adminuser";
+    public const string UserGroups = "usergroups";
+    public const string RolePermissionInfo = "rolepermissioninfo";
+    public const string IdentityProviders = "identityproviders";
+
+  }
+
+  public static class OrganisationHeaderMap
+  {
+    public const string Identifier_Id = "identifier_id";
+    public const string Identifier_LegalName = "identifier_legalname";
+    public const string Identifier_Uri = "identifier_uri";
+    public const string Identifier_Scheme = "identifier_scheme";
+    public const string AdditionalIdentifiers = "additionalIdentifiers";
+    public const string Address_StreetAddress = "address_streetaddress";
+    public const string Address_Locality = "address_locality";
+    public const string Address_Region = "address_region";
+    public const string Address_PostalCode = "address_postalcode";
+    public const string Address_CountryCode = "address_countrycode";
+    public const string Address_CountryName = "address_countryname";
+    public const string Detail_Organisation_Id = "detail_organisation_id";
+    public const string Detail_CreationDate = "detail_creationdate";
+    public const string Detail_BusinessType = "detail_businesstype";
+    public const string Detail_SupplierBuyerType = "detail_supplierbuyertype";
+    public const string Detail_IsSme = "detail_is_sme";
+    public const string Detail_IsVcse = "detail_is_vcse";
+    public const string Detail_RightToBuy = "detail_rightTobuy";
+    public const string Detail_IsActive = "detail_isactive";
+
+    public const string AdditionalIdentifiers_Id = "Id";
+    public const string AdditionalIdentifiers_LegalName = "LegalName";
+    public const string AdditionalIdentifiers_URI = "Uri";
+    public const string AdditionalIdentifiers_Scheme = "Scheme";
+
+  }
+
+  public static class ContactsHeaderMap
+  {
+    public const string ContactType = "contact_type";
+    public const string ContactID = "contact_id";
+    public const string ContactPointID = "contactpoint_id";
+    public const string OriginalContactPointID = "original_contactpoint_id";
+    public const string AssignedContactType = "assigned_contact_type";
+    public const string Contact_ContactID = "contacts_contactid";
+    public const string Contacts_ContactType = "contacts_contacttype";
+    public const string Contacts_ContactValue = "contacts_contactvalue";
+    public const string ContactPoint_Reason = "contactpoint_reason";
+    public const string ContactPoint_Name = "contactpoint_name";
+  }
+
+  public static class AuditLogHeaderMap
+  {
+    public const string ID = "ID";
+    public const string Event = "Event";
+    public const string UserId = "UserId";
+    public const string Application = "Application";
+    public const string ReferenceData = "ReferenceData";
+    public const string IpAddress = "IpAddress";
+    public const string Device = "Device";
+  }
 }
